First working ascii parser for pcd files
This is a first implementation and there is probably a lot to improve
here. The parsers are now implemented via macros as it cleans up the
code a lot by allowing static parsers for each type. Different data
types inside points are not handled for now, but the macros might allow
to implement this more easily.

Added example pcd files for future use. The files are from the PCL
project and their documentation, thanks for that.

Added pretty-assertions as a dev dependency. These assertions allow for
better error detection while testing large pcd files.