Reworked pcd::Schema and pcd::Error; Started example ascii parser
Schema now uses a HashMap with the identifier/name as the key eg. ("x",
"rgba", etc.) and the FieldDef with the type and count as the value of
the map. Extended the pcd::Error type to include more errors mainly for
schema mismatches and io/parsing errors. Started example ascii parser
for PointXYZRGBA using PcdDeserialize.
Hopefully this can be translated to a derive macro, to easily implement
this for all point types.